--- examples/langserve/.env
@@ -1,2 +1,0 @@
-OPENAI_API_KEY=
-AGENTWALLET_API_KEY=
