+++ examples/langserve/.env.example
@@ -1,0 +1,2 @@
+OPENAI_API_KEY=
+AGENTWALLET_API_KEY=
